requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.2
tensorflow-cpu==2.21.0
keras==3.13.2
numpy==2.1.3
Pillow==11.3.0
python-multipart==0.0.20
